requirements.txt
django>=5.0,<6.0
djangorestframework>=3.16,<4.0
requests>=2.32,<3.0
Pillow>=12.0,<13.0

